refactor: Move server into index.js, update script cleanup
