+++ agent-streaming/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --env-file=.env --watch src/server.js",
-    "dev": "node --env-file=.env --watch src/server.js"
+    "start": "node --env-file=.env --watch src/index.js",
+    "dev": "node --env-file=.env --watch src/index.js"
   },
   "keywords": [],
   "author": "",
